+++ test/widget_test.dart
@@ -1,0 +1,8 @@
+// Copyright 2022 Andrious Solutions Ltd. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+/// Calls the example app's very own test
+import '../example/test/widget_test.dart';
+
+void main() => testMyApp();